+++ src/main/java/com/microsoft/graph/requests/SecurityRequestBuilder.java
@@ -58,6 +58,26 @@ public class SecurityRequestBuilder extends BaseRequestBuilder<Security> {
     }
 
 
+    /**
+     *  Gets a request builder for the SubjectRightsRequest collection
+     *
+     * @return the collection request builder
+     */
+    @Nonnull
+    public com.microsoft.graph.requests.SubjectRightsRequestCollectionRequestBuilder subjectRightsRequests() {
+        return new com.microsoft.graph.requests.SubjectRightsRequestCollectionRequestBuilder(getRequestUrlWithAdditionalSegment("subjectRightsRequests"), getClient(), null);
+    }
+
+    /**
+     * Gets a request builder for the SubjectRightsRequest item
+     *
+     * @return the request builder
+     * @param id the item identifier
+     */
+    @Nonnull
+    public com.microsoft.graph.requests.SubjectRightsRequestRequestBuilder subjectRightsRequests(@Nonnull final String id) {
+        return new com.microsoft.graph.requests.SubjectRightsRequestRequestBuilder(getRequestUrlWithAdditionalSegment("subjectRightsRequests") + "/" + id, getClient(), null);
+    }
 
     /**
      * Gets the request builder for CasesRoot