+++ src/main/java/com/microsoft/graph/requests/SubjectRightsRequestCollectionRequest.java
@@ -7,7 +7,7 @@ package com.microsoft.graph.requests;
 
 import com.microsoft.graph.http.IRequestBuilder;
 import com.microsoft.graph.core.ClientException;
-import com.microsoft.graph.models.Privacy;
+import com.microsoft.graph.models.Security;
 import com.microsoft.graph.models.SubjectRightsRequest;
 import java.util.Arrays;
 import java.util.EnumSet;